pyproject.toml
[tool.poetry]
name = "papers-fetcher"
version = "0.1.0"
description = "Fetch PubMed papers with company affiliations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.0"
biopython = "^1.81"
beautifulsoup4 = "^4.12.0"
python-dotenv = "^1.0.0"  # Added for environment variable support

[tool.poetry.scripts]
get-papers-list = "cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"